preliminary next/prev month buttons, refactor and prepare for more tests
